GUI Note
Use Platform.runLater when you are executing off of the JavaFX Application Thread and you want to run some logic on the JavaFX application Thread.
Use a Worker when you are running on the JavaFX Application Thread and want to spawn some logic or (especially) I/O on a new thread so that you don't block the JavaFX Application Thread.
You would never want to do network I/O inside a Platform.runLater's run method, but would often want to do it in a Worker's call method. 
(StackOverFlow)